Update #2
Added A LOT of UI assets. Switched to Text Mesh Pro. Added functional coding for different information UI elements to activate, including Intel. Also adjusted UI in relation to different screen variations. Added tags for different years, and the rest of the countries to be able to be explored. Also changed, "Send Agents to" rather than "Travel to" as I think that makes more sense.