Update #9
Yay! I didn't lie this time. This is probably going to be the biggest update for the game. Probably... Added all the scenes for each month/year, and End Scene (Empty currently). Raycasting Script has been updated with each location and interactions, as well as dates. It's all functional now! Updated particle effects to reflect occupation of countries. Well, it's not in the code, but I also figured out how I want to go about the story now and how information will be given! It's in a Google Document, I may link it to the GitHub. Next Update so long as nothing breaks will probably be the big Information update, as well an update to the 'chances' of being caught. Then I'll be really close to being done! :-D