Update #13
Welp, as it turns out, I can't think or function when I'm this tired. I'm going to take a power nap and come back to it in the morning. There is just no feasible way for me at this point to get it done this night unfortunately. In this update, I added a little over half of the statuses of the countries. Things I have to do left is to add the rest of the statuses, the successful reports, and the end scene... Oooh boy. Probably should have asked for an extension for the project and the video rather than just the video... :(