Update #15
Added all Reports for Countries! All that's left is to create and populate the End scene. Not much left! I also realize that the random numbers could probably be lessened the chance of getting caught for Medium and High, but as of now, I don't got the time for this...