Update #16 - Final for Project
Well, I've implemented an EndScene that doesn't work in the end as far as I can tell unfortunately. But, at this point, I've run out of steam for this project. I'll probably come back to it to polish it after this class, but until then, it is what it is.